refactor: modify base class
