+++ eventhandlers/action_runners/ExportActionRunner.py
@@ -1,4 +1,4 @@
-from action_runners.ActionRunner import ActionRunner
+from action_runners.base.ActionRunner import ActionRunner
 from shared.shared_logger import get_shared_logger
 from shared.regular.regular_log import log_has_error
 from shared.export.export_create import create_new_export
