+++ eventhandlers/action_runners/SampleActionBoilerplate.py
@@ -11,10 +11,16 @@ class MySampleAction(ActionRunner):
     category = 'some_category'  # Optional
 
     # What events can this action listen to?
-    trigger = ActionTrigger(default_event = 'some_diffgram_event', event_list = ['some_diffgram_event'])
+    trigger = ActionTrigger(default_event = 'some_diffgram_event',
+                            event_list = ['some_diffgram_event'])
 
-    condition_data = ActionCondition()  # What pre-conditions can this action have?
-    completion_condition_data = ActionCompleteCondition()  # How to declare the actions as completed?
+    # What pre-conditions can this action have?
+    condition_data = ActionCondition(default_event = 'some_diffgram_event',
+                                     event_list = ['some_diffgram_event'])
+
+    # How to declare the actions as completed?
+    completion_condition_data = ActionCompleteCondition(default_event = 'some_diffgram_event',
+                                                        event_list = ['some_diffgram_event'])
 
     def execute_pre_conditions(self, session) -> bool:
         # Return true if no pre-conditions are needed.
